organize auth concept files
