+++ .gitignore
@@ -45,4 +45,5 @@ package-lock.json
 .sentryclirc
 
 # cypress/e2e/1-getting-started
-# cypress/e2e/2-advanced-examples
+# cypress/e3e/2-advanced-examples
+screenshots
