+++ src/concepts/Auth/repository/fetchCurrentUserRequest.ts
@@ -1,4 +1,4 @@
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api"
 
 export const fetchCurrentUserRequest = () => {
     return api.get("/user")
