+++ src/concepts/Auth/views/Register/Register.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/breakpoints.scss";
+@import "../../../../components/theme/breakpoints.scss";;
 
 .content * {
   color: var(--text-primary);
