--- src/concepts/auth/index.ts
@@ -1,2 +1,0 @@
-export { AuthProvider } from "./view/components/AuthProvider";
-export * from "./view/hooks/useAuth";
