+++ src/concepts/infra/http/createBudgetRequest.ts
@@ -1,4 +1,4 @@
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const createBudgetRequest = async ({ name }: {name:string}) => {
   return api.post("/budgets", {
