+++ src/concepts/infra/http/createTransactionRequest.ts
@@ -1,5 +1,5 @@
 import { NewTransaction } from "@/concepts/transactions/domain/Transaction";
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const createTransactionRequest = async (transaction: NewTransaction) => {
   return api.post("/transactions", {
