+++ src/concepts/infra/http/deleteBudgetRequest.ts
@@ -1,4 +1,4 @@
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const deleteBudgetRequest = async (id: number) => {
   return api.delete(`/budgets/${id}`);
