+++ src/concepts/infra/http/deleteTransactionRequest.ts
@@ -1,4 +1,4 @@
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 type DeleteTransactionParams = {
   id: number,
