+++ src/concepts/infra/http/fetchBudgetsRequest.ts
@@ -1,4 +1,4 @@
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const fetchBudgetsRequest = async () => {
   return api.get(`/budgets`);
