+++ src/concepts/infra/http/updateBudgetRequest.ts
@@ -1,5 +1,5 @@
 import { Budget } from "@/concepts/transactions/domain/Budget";
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const updateBudgetRequest = async ({ id, name }: Budget) => {
   return api.put(`/budgets/${id}`, {
