+++ src/concepts/infra/http/updateTransactionRequest.ts
@@ -1,5 +1,5 @@
 import { Transaction } from "@/concepts/transactions/domain/Transaction";
-import { api } from "../services/apiClient";
+import { api } from "@/lib/http/api";
 
 export const updateTransactionRequest = async (transaction: Transaction) => {
   return api.put(`/transactions/${transaction.id}`, {
