--- src/concepts/infra/services/apiClient.ts
@@ -1,3 +1,0 @@
-import { setupApiClient } from "./api";
-
-export const api = setupApiClient();
