+++ src/concepts/pages/home/Home.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/breakpoints.scss";
+@import "../../../components/theme/breakpoints.scss";;
 
 .container {
   padding: 12px;
