+++ src/concepts/pages/transactions/components/TransactionsList.module.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/breakpoints.scss";
+@import "../../../../components/theme/breakpoints.scss";
 
 .container {
   display: grid;
