+++ src/concepts/ui/Button/Button.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/breakpoints.scss";
+@import "../../../components/theme/breakpoints.scss";;
 
 .button {
   width: 100%;
