+++ src/concepts/ui/ErrorMessage/ErrorMessage.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/breakpoints.scss";
+@import "../../../components/theme/breakpoints.scss";;
 
 li.error {
   color: #be6e6e;
