+++ src/concepts/ui/Header/Header.module.scss
@@ -1,5 +1,4 @@
-@import "../../styles/breakpoints.scss";
-
+@import "../../../components/theme/breakpoints.scss";;
 
 .active * {
   color: black;
