+++ src/concepts/ui/Layout/Layout.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/breakpoints.scss";
+@import "../../../components/theme/breakpoints.scss";;
 
 .layout {
   display: flex;
