+++ src/lib/http/errors/AuthTokenError.ts
