+++ src/pages/_app.tsx
@@ -1,12 +1,12 @@
 import type { AppProps } from "next/app";
-import "@/concepts/styles/global.scss";
+import "@/components/theme/global.scss";
 import { Layout } from "@/concepts/ui/Layout/Layout";
-import { AuthProvider } from "@/concepts/auth";
 import { TransactionProvider, BudgetProvider } from "@/concepts/transactions";
 import { ConfirmProvider } from "material-ui-confirm";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { ToastContainer } from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css';
+import { AuthProvider } from "@/concepts/Auth/store/AuthProvider";
 
 const queryClient = new QueryClient()
