+++ src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { withSSRAuth } from "@/concepts/auth/middlewares/withSSRAuth";
+import { withSSRAuth } from "@/concepts/Auth/middlewares/withSSRAuth";
 import { HomePage } from "@/concepts/pages/home/HomePage";
 
 const Page = () => {
